Moved render functions to separate file in utils, updated imports
