+++ src/components/tasks.js
@@ -1,4 +1,4 @@
-import {createNode} from '../utils';
+import {createNode} from '../utils/render';
 
 const createTasksTemplate = () => {
   return (
