+++ src/utils/index.js
@@ -1,5 +1,3 @@
-import {RENDER_POSITION} from '../const';
-
 const castTimeFormat = (value) => {
   return value < 10 ? `0${value}` : String(value);
 };
@@ -9,23 +7,6 @@ export const formatTime = (date) => {
   const minutes = castTimeFormat(date.getMinutes());
 
   return `${hours}:${minutes}`;
-};
-
-export const createNode = (template) => {
-  const node = document.createElement(`div`);
-  node.innerHTML = template;
-  return node.firstChild;
-};
-
-export const render = (container, element, location = RENDER_POSITION.BEFOREEND) => {
-  switch (location) {
-    case RENDER_POSITION.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RENDER_POSITION.BEFOREEND:
-      container.append(element);
-      break;
-  }
 };
 
 export const getRanodomBoolean = () => Math.random() > 0.5;
